+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Features
 
 * Add `row_click` option for `Cinder.Table.table`, to make entire rows clickable
+* Support `scope` and `tenant` options to `Cinder.Table.table`
+  * `tenant` can also be passed in as part of the `query_opts` option
 
 ### Bug fixes